Refactor combat stats and mainloop
First, we change `CombatResult` into `CombatStats`, which serves as a
structure for both pre-compbat stats and post-combat results. It also
inherit the responsibility of making the actual combat rolls. We thus
remove the `attack()` function.

Then, we introduce a new `MainloopState` enum to track global mainloop
state. This will make popup dialog management much clearer, especially
when we add combat confirmation dialog in the upcoming commit.