Major refactoring of the screen drawing code
We introduce `HexCell`, which is a sub-surface we draw the hex cell's
content in before sending it onto the screen.

For now (because the refactoring as it is is big enough for this
commit), we have a  single hex cell used to draw all cells during the
main `draw()` phase, but eventually, they'll be lock to a relative
position and we'll maintain one cell per "hex-in-screen" and it should
make drawing more straightforward.

This change was necessary to facilitate the coming of "cell highlight"
capabilities which I first want to bring on with a "move to" feature
(`m` key in civ5).