Add the new Move Mode!
It works like the `m` key in civ 5. This was some serious fun to
implement!

There's a bit of code duplication between the "normal move" and "move
move" code right now, it's going to gradually go away as attacking is
made possible through this new mode (the challenge is to make sure that
there's a free cell next to the target before considering the enemy
attackable).