Have one persistent HexCell per visible cell
This replaces the old `VisiblePosIterator` logic which iterated over all
visible pos, redrawing with a single `HexCell` iteratively.

This new approach is simpler and will allow for finer grained drawing
call from `Game` to `Screen`. With the old approach, the cell rendering
was not kept, so, for example, it wasn't possible, from `Game`, to call
`draw()` and then highlight some cells. It is now possible.